no-glib: remove glib error support.
Most of the g_set_error's is not very useful. And there're many plugins
simply ignore the error parameter, Now errors are logged on-site. Only
problem is that some g_set_error is used for returning errors to the
client.

To solve that, now detailed error messages are no longer returned to
the client, but logged. The client only receive an abstract error.
Actually this approach is not that different from what mpd does now.

Now mpd compiles and works, but may contain bugs, and some plugins may
not compile.